Updated site.

+++ build_with_commit.sh
@@ -4,31 +4,31 @@
 bundle exec jekyll build
 
 sass assets/css/main.scss:assets/css/main.css
-rm -rf .sass-cache*
+rm -rf .sass-cache
 source build_config.sh
 
 # goto destination - delete all apart from git.
 
-if [ -d "$dest" ]; then
+if [ -d ${dest} ] && [ -n ${1} ]; then
     echo $dest
 
     # Move to the output directory
     cd $dest
 
-    # Remove everything apart from the git
+    # Remove everything from $dest apart from the .git folder
     ls -a | grep -v ".git" | tail -n +3 |xargs rm -rf
 
 
     # Move back to the original directory.
     cd -
 
-    # Now back in the original directory.
+    # Now, back in the original directory.
     # Copy site output to destination
     cp -r  ./_site/* $dest
 
+
     cd $dest
     git add -A
-
     git commit -m "Updated site. $1"
